Challenge List View 수정
