+++ lib/main.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:provider/provider.dart';
 import 'package:yum_application/app_config.dart';
-import 'package:yum_application/src/ingredient/view/challenge_detail_view.dart';
-import 'package:yum_application/src/ingredient/view/challenge_list_view.dart';
+import 'package:yum_application/src/challege/view/challenge_detail_view.dart';
+import 'package:yum_application/src/challege/view/challenge_list_view.dart';
 import 'package:yum_application/src/ingredient/view/home_view.dart';
 import 'package:yum_application/src/util/app_theme.dart';
 import 'package:yum_application/src/util/global_variable.dart';
